services/inference/requirements.txt
python-doctr
Pillow
numpy
